First outline push
This is the first outline of the working scripts. The main wrapper pulls two main functions from tools. One function to import a csv/tsv and one function to query the API. Functions are split into utilities (utils.py) and visualizations (figs.py). Please add accordingly. *don't forget to add to the init!*